+++ js/app-config.js
@@ -1,4 +1,4 @@
 window.UmmibyAppConfig = Object.freeze({
   name: "Ummiby Companion",
-  version: "0.6.5.2"
+  version: "0.6.6.1"
 });